llm-flask/embed/requirements.txt
google-cloud-vision
faiss-cpu
sentence-transformers
numpy
google-cloud-storage
pymupdf
-e ../config